src/my_voice_assistant/requirements.txt
gtts==2.5.4
openai==1.81.0
pyaudio==0.2.14
python-dotenv==1.1.0
setuptools==75.8.0
vosk==0.3.45
websocket_client==1.8.0
tenacity
pygame
# RAG dependencies
chromadb==0.4.24
sentence-transformers==2.7.0
langchain==0.2.11
langchain-community==0.2.10
langchain-openai==0.1.19
faiss-cpu==1.8.0
numpy==1.24.3
tiktoken==0.7.0
